separate old thread gql tests from new api tests
